refactor: clean up e2e framework job utils
1. Move job_util to e2e/framework/job package
2. Remove unused UpdateJobWithFunc
