+++ test/e2e/node/ttlafterfinished.go
@@ -24,6 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/kubernetes/pkg/util/slice"
 	"k8s.io/kubernetes/test/e2e/framework"
+	jobutil "k8s.io/kubernetes/test/e2e/framework/job"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -47,11 +48,11 @@ func cleanupJob(f *framework.Framework, job *batch.Job) {
 	removeFinalizerFunc := func(j *batch.Job) {
 		j.ObjectMeta.Finalizers = slice.RemoveString(j.ObjectMeta.Finalizers, dummyFinalizer, nil)
 	}
-	_, err := framework.UpdateJobWithRetries(c, ns, job.Name, removeFinalizerFunc)
+	_, err := jobutil.UpdateJobWithRetries(c, ns, job.Name, removeFinalizerFunc)
 	Expect(err).NotTo(HaveOccurred())
-	framework.WaitForJobGone(c, ns, job.Name, wait.ForeverTestTimeout)
+	jobutil.WaitForJobGone(c, ns, job.Name, wait.ForeverTestTimeout)
 
-	err = framework.WaitForAllJobPodsGone(c, ns, job.Name)
+	err = jobutil.WaitForAllJobPodsGone(c, ns, job.Name)
 	Expect(err).NotTo(HaveOccurred())
 }
 
@@ -64,27 +65,27 @@ func testFinishedJob(f *framework.Framework) {
 	backoffLimit := int32(2)
 	ttl := int32(10)
 
-	job := framework.NewTestJob("randomlySucceedOrFail", "rand-non-local", v1.RestartPolicyNever, parallelism, completions, nil, backoffLimit)
+	job := jobutil.NewTestJob("randomlySucceedOrFail", "rand-non-local", v1.RestartPolicyNever, parallelism, completions, nil, backoffLimit)
 	job.Spec.TTLSecondsAfterFinished = &ttl
 	job.ObjectMeta.Finalizers = []string{dummyFinalizer}
 	defer cleanupJob(f, job)
 
 	framework.Logf("Create a Job %s/%s with TTL", ns, job.Name)
-	job, err := framework.CreateJob(c, ns, job)
+	job, err := jobutil.CreateJob(c, ns, job)
 	Expect(err).NotTo(HaveOccurred())
 
 	framework.Logf("Wait for the Job to finish")
-	err = framework.WaitForJobFinish(c, ns, job.Name)
+	err = jobutil.WaitForJobFinish(c, ns, job.Name)
 	Expect(err).NotTo(HaveOccurred())
 
 	framework.Logf("Wait for TTL after finished controller to delete the Job")
-	err = framework.WaitForJobDeleting(c, ns, job.Name)
+	err = jobutil.WaitForJobDeleting(c, ns, job.Name)
 	Expect(err).NotTo(HaveOccurred())
 
 	framework.Logf("Check Job's deletionTimestamp and compare with the time when the Job finished")
-	job, err = framework.GetJob(c, ns, job.Name)
+	job, err = jobutil.GetJob(c, ns, job.Name)
 	Expect(err).NotTo(HaveOccurred())
-	finishTime := framework.JobFinishTime(job)
+	finishTime := jobutil.FinishTime(job)
 	finishTimeUTC := finishTime.UTC()
 	Expect(finishTime.IsZero()).NotTo(BeTrue())
