+++ test/e2e/upgrades/nvidia-gpu.go
@@ -24,6 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/kubernetes/test/e2e/framework"
 	"k8s.io/kubernetes/test/e2e/framework/gpu"
+	jobutil "k8s.io/kubernetes/test/e2e/framework/job"
 	"k8s.io/kubernetes/test/e2e/scheduling"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 
@@ -54,7 +55,7 @@ func (t *NvidiaGPUUpgradeTest) Test(f *framework.Framework, done <-chan struct{}
 	t.verifyJobPodSuccess(f)
 	if upgrade == MasterUpgrade || upgrade == ClusterUpgrade {
 		// MasterUpgrade should be totally hitless.
-		job, err := framework.GetJob(f.ClientSet, f.Namespace.Name, "cuda-add")
+		job, err := jobutil.GetJob(f.ClientSet, f.Namespace.Name, "cuda-add")
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 		gomega.Expect(job.Status.Failed).To(gomega.BeZero(), "Job pods failed during master upgrade: %v", job.Status.Failed)
 	}
@@ -69,7 +70,7 @@ func (t *NvidiaGPUUpgradeTest) Teardown(f *framework.Framework) {
 func (t *NvidiaGPUUpgradeTest) startJob(f *framework.Framework) {
 	var activeSeconds int64 = 3600
 	// Specifies 100 completions to make sure the job life spans across the upgrade.
-	testJob := framework.NewTestJob("succeed", "cuda-add", v1.RestartPolicyAlways, 1, 100, &activeSeconds, 6)
+	testJob := jobutil.NewTestJob("succeed", "cuda-add", v1.RestartPolicyAlways, 1, 100, &activeSeconds, 6)
 	testJob.Spec.Template.Spec = v1.PodSpec{
 		RestartPolicy: v1.RestartPolicyOnFailure,
 		Containers: []v1.Container{
@@ -86,11 +87,11 @@ func (t *NvidiaGPUUpgradeTest) startJob(f *framework.Framework) {
 		},
 	}
 	ns := f.Namespace.Name
-	_, err := framework.CreateJob(f.ClientSet, ns, testJob)
+	_, err := jobutil.CreateJob(f.ClientSet, ns, testJob)
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	framework.Logf("Created job %v", testJob)
 	ginkgo.By("Waiting for gpu job pod start")
-	err = framework.WaitForAllJobPodsRunning(f.ClientSet, ns, testJob.Name, 1)
+	err = jobutil.WaitForAllJobPodsRunning(f.ClientSet, ns, testJob.Name, 1)
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	ginkgo.By("Done with gpu job pod start")
 }
@@ -99,9 +100,9 @@ func (t *NvidiaGPUUpgradeTest) startJob(f *framework.Framework) {
 func (t *NvidiaGPUUpgradeTest) verifyJobPodSuccess(f *framework.Framework) {
 	// Wait for client pod to complete.
 	ns := f.Namespace.Name
-	err := framework.WaitForAllJobPodsRunning(f.ClientSet, f.Namespace.Name, "cuda-add", 1)
+	err := jobutil.WaitForAllJobPodsRunning(f.ClientSet, f.Namespace.Name, "cuda-add", 1)
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	pods, err := framework.GetJobPods(f.ClientSet, f.Namespace.Name, "cuda-add")
+	pods, err := jobutil.GetJobPods(f.ClientSet, f.Namespace.Name, "cuda-add")
 	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	createdPod := pods.Items[0].Name
 	framework.Logf("Created pod %v", createdPod)
